// src/common/error-codes.enum.ts
export enum ErrorCode {
  VALIDATION_ERROR = "VALIDATION_ERROR",
  MISSING_REQUIRED_FIELD = "MISSING_REQUIRED_FIELD",

  VM_NOT_FOUND = "VM_NOT_FOUND",
  VM_CREATION_FAILED = "VM_CREATION_FAILED",
  VM_DELETION_FAILED = "VM_DELETION_FAILED",
  VM_STATE_FETCH_FAILED = "VM_STATE_FETCH_FAILED",

  TASK_NOT_FOUND = "TASK_NOT_FOUND",

  PULUMI_STACK_ERROR = "PULUMI_STACK_ERROR",

  DATABASE_ERROR = "DATABASE_ERROR",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",

  VM_START_FAILED = "VM_START_FAILED",
  VM_STOP_FAILED = "VM_STOP_FAILED",
  VM_RESTART_FAILED = "VM_RESTART_FAILED",
  VM_ALREADY_STARTED = "VM_ALREADY_STARTED",
  VM_ALREADY_STOPPED = "VM_ALREADY_STOPPED",
  PERMISSION_DENIED = "PERMISSION_DENIED",
  HYPERVISOR_OFFLINE = "HYPERVISOR_OFFLINE",
}

export const ErrorCodeToStatusMap: Record<ErrorCode, number>= {
  [ErrorCode.VALIDATION_ERROR]: 400,
  [ErrorCode.MISSING_REQUIRED_FIELD]: 400,
  [ErrorCode.VM_NOT_FOUND]: 404,
  [ErrorCode.TASK_NOT_FOUND]: 404,
  [ErrorCode.VM_CREATION_FAILED]: 500,
  [ErrorCode.VM_DELETION_FAILED]: 500,
  [ErrorCode.VM_STATE_FETCH_FAILED]: 500,
  [ErrorCode.PULUMI_STACK_ERROR]: 500,
  [ErrorCode.DATABASE_ERROR]: 500,
  [ErrorCode.INTERNAL_SERVER_ERROR]: 500,
  [ErrorCode.VM_START_FAILED]: 500,
  [ErrorCode.VM_STOP_FAILED]: 500,
  [ErrorCode.VM_RESTART_FAILED]: 500,
  [ErrorCode.VM_ALREADY_STARTED]: 409,
  [ErrorCode.VM_ALREADY_STOPPED]: 409,
  [ErrorCode.PERMISSION_DENIED]: 403,
  [ErrorCode.HYPERVISOR_OFFLINE]: 503,
};